move password_reader_mock.go to mock testonly package
