+++ cmd/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "flags.go",
         "helpers.go",
         "password_reader.go",
-        "password_reader_mock.go",
         "wrap_flags.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/cmd",
@@ -16,7 +15,6 @@ go_library(
     deps = [
         "//config/params:go_default_library",
         "//io/file:go_default_library",
-        "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
@@ -35,6 +33,7 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
+        "//cmd/mock:go_default_library",
         "//config/params:go_default_library",
         "//testing/assert:go_default_library",
         "//testing/require:go_default_library",
